Fix for visual studio tests
A CLASSIC. Why this stupid testing framework cannot cast from int to size_t, why

+++ src/materials/material_library.cpp
@@ -60,7 +60,7 @@ const Bsdf* MaterialLibrary::get_default() const
     return MaterialLibrary::default_material;
 }
 
-int MaterialLibrary::size() const
+size_t MaterialLibrary::size() const
 {
-    return (int)lib.size();
+    return lib.size();
 }
